made the viewmodel and fragment for publish settings abstract
